requirements.txt
asgiref==3.2.10
certifi==2020.6.20
Django==3.0.8
django-cors-headers==3.2.1
django-elasticsearch-dsl==7.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
elasticsearch==7.8.0
elasticsearch-dsl==7.2.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
